Create Donate Page
This page pops open when a user clicks on the "Donate" option. The page allows the user to view all previously donated items. Users can also update posted items by clicking on the edit or delete option. If a user wishes to donate more items, the user will simply click on the "add item" option at the top right of the page.