Request Form
Once a user indicates interest in a posted item, the user will be redirected to the Request Form. Here the user is required to fill in a username, school email address, item ID, and message to the donor (optional). Afterward, the user would click submit and await a response from the donor at the user's school email address