Create View History
On the web application, users will be allowed to view the history of items they have donated or received. To view the history of received items, users should click on the "received" button. To view the history of a donated item, users should click on the Donated button; these two buttons are located at the top of the page, while the history of items is listed chronologically on the page.