Create Donation Approval Page
To successfully add or edit an item, the user will be required to fill out an item donation approval form, which will be approved or declined by the admin. The user will be required to enter the item name, upload picture(s) of the item, and write a short description of the item. After completing the form, the user is required to click on "submit."