Create Login
There are two buttons on the application that will allow login or registration and a link to help retrieve a password. From the above, the two buttons will be linked to the Firebase server and connected to the login and sign-up forms. Only users with a student or FUPRE email address will be allowed to login and successfully register in the web application.